remove manual torch seeds from tests. formatting.
One particular test for inference accuracy based on MMD was failing
unrelated to the changes in this PR. This is fixed with this commit
by removing the manual seeding in some of the test files. This will
be fixed in a separate PR. For details see issue #130.

THis commits contains formatting changes for some other files.