require x_shape in NeuralPosterior.
It is reasonable to pass an x_shape because the
posterior will eventually be conditioned and in
all occurrences in sbi where instantiate the
posterior x_shape can be inferred. Alternatively,
one would have raise a warning if it is None and
infer it from simulating once when sample, log_prob,
or default_x is called.